+++ src/pages/Aademy/KidsTutorial/assets/images/left-kid.png
